backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
replicate==1.0.7
opencv-python-headless==4.12.0.88
pillow==12.1.0
numpy==2.2.6
python-dotenv==1.2.1
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
httpx==0.28.1

